# app/swagger/swagger.yaml
basePath: /
definitions:
...
  node.ActionRule:
    properties:
      actType:
        description: and 表示会签 or表示或签，默认为或签
        type: string
      autoUp:
        type: boolean
      isEmpty:
        type: boolean
      labelNames:
        type: string
      labels:
        type: integer
      level:
        type: integer
      memberCount:
        description: 表示需要通过的人数 如果是会签
        type: integer
      type:
        type: string
    type: object
  node.CondNode:
    properties:
      attrs:
        $ref: '#/definitions/node.UserNode'
      type:
        type: string
      value:
        type: string
    type: object
  node.ConditionNode:
    properties:
      boundEqual:
        type: string
      conditions:
        items:
          $ref: '#/definitions/node.CondNode'
        type: array
      isEmpty:
        type: boolean
      lowerBound:
        description: 类型为range
        type: string
      lowerBoundEqual:
        type: string
      oriValue:
        items:
          type: string
        type: array
      paramKey:
        type: string
      paramLabel:
        type: string
      paramValues:
        description: 类型为 value
        items:
          type: string
        type: array
      type:
        type: string
      unit:
        type: string
      upperBound:
        type: string
      upperBoundEqual:
        type: string
    type: object
  node.Node:
    properties:
      childNode:
        $ref: '#/definitions/node.Node'
      conditionNodes:
        items:
          $ref: '#/definitions/node.Node'
        type: array
      name:
        type: string
      nodeId:
        type: string
      prevId:
        type: string
      properties:
        $ref: '#/definitions/node.Properties'
      type:
        type: string
    type: object
  node.Properties:
    properties:
      actionRules:
        items:
          $ref: '#/definitions/node.ActionRule'
        type: array
      activateType:
        description: ONE_BY_ONE 代表依次审批
        type: string
      agreeAll:
        type: boolean
      conditions:
        items:
          items:
            $ref: '#/definitions/node.ConditionNode'
          type: array
        type: array
      noneActionAction:
        type: string
      script:
        type: string
    type: object
  node.UserNode:
    properties:
      avatar:
        type: string
      name:
        type: string
    type: object
...
  request.ProcRequest:
    properties:
      company:
        description: 用户所在公司
        type: string
      name:
        type: string
      resource:
        $ref: '#/definitions/node.Node'
        description: 流程定义json字符串
      userId:
        description: 用户id
        type: string
      username:
        type: string
    type: object
  request.ReceiverProcess:
    properties:
      company:
        type: string
      department:
        type: string
      procInstID:
        type: string
      procName:
        type: string
      title:
        type: string
      userId:
        type: string
      username:
        type: string
      var:
        additionalProperties:
          type: string
        type: object
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  workflow.ProcPageReceiver:
    properties:
      company:
        description: 用户所在公司
        type: string
      created_at:
        type: string
      desc:
        description: 排序方式:升序false(默认)|降序true
        type: boolean
      filter:
        items:
          type: string
        type: array
      id:
        description: 主键
        type: integer
      name:
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      resource:
        description: 流程定义json字符串
        type: string
      updated_at:
        type: string
      userId:
        description: 用户id
        type: string
      username:
        type: string
      version:
        type: integer
    type: object
  workflow.ProcessPageReceiver:
    properties:
      candidate:
        description: 审批人
        type: string
      company:
        type: string
      created_at:
        type: string
      department:
        description: 用户部门
        type: string
      desc:
        description: 排序方式:升序false(默认)|降序true
        type: boolean
      duration:
        type: integer
      endTime:
        type: string
      filter:
        items:
          type: string
        type: array
      id:
        description: 主键
        type: integer
      isFinished:
        type: boolean
      nodeID:
        description: 当前节点
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      procDefId:
        description: 流程定义ID
        type: integer
      procDefName:
        description: 流程定义名
        type: string
      startUserId:
        type: string
      startUserName:
        type: string
      taskID:
        description: 当前任务
        type: integer
      title:
        description: title 标题
        type: string
      updated_at:
        type: string
    type: object
  workflow.TaskReceiver:
    properties:
      candidate:
        type: string
      comment:
        type: string
      company:
        type: string
      pass:
        type: string
      procInstID:
        type: integer
      taskID:
        type: integer
      userID:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: This is Devops API
  title: Swagger Devops API
  version: 0.0.1
paths:
...
  /v1/workflow/identity/{id}:
    post:
      consumes:
      - application/json
      description: 获取流程的所有参与者
      parameters:
      - description: 流程id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取流程的所有参与者
      tags:
      - process
  /v1/workflow/proc/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除工作流接口
      parameters:
      - description: 工作流id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除工作流接口
      tags:
      - proc
  /v1/workflow/proc/list:
    post:
      consumes:
      - application/json
      description: 查询定义的工作流列表
      parameters:
      - description: page  pageSize  filter 筛选条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/workflow.ProcPageReceiver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 查询定义的工作流列表
      tags:
      - proc
  /v1/workflow/proc/save:
    post:
      consumes:
      - application/json
      description: 定义工作流接口
      parameters:
      - description: 流程名， 用户id， 用户所在的公司，定义流程json串
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ProcRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 定义工作流接口
      tags:
      - proc
  /v1/workflow/process/inst:
    post:
      consumes:
      - application/json
      description: 查询到我审批的流程实例
      parameters:
      - description: 查询轮到自己审批的流程
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/workflow.ProcessPageReceiver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 查询到我审批的流程实例
      tags:
      - process
  /v1/workflow/process/inst/myself:
    post:
      consumes:
      - application/json
      description: 查询到我发起的流程实例
      parameters:
      - description: 查询到我发起的流程实例
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/workflow.ProcessPageReceiver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 查询到我发起的流程实例
      tags:
      - process
  /v1/workflow/process/inst/notify:
    post:
      consumes:
      - application/json
      description: 查询到我发起的流程实例
      parameters:
      - description: 查询到我发起的流程实例
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/workflow.ProcessPageReceiver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 查询抄送我的流程实例
      tags:
      - process
  /v1/workflow/process/start:
    post:
      consumes:
      - application/json
      description: 开启工作流实例
      parameters:
      - description: 模型ReceiverProcess 中的数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReceiverProcess'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 开启工作流实例
      tags:
      - process
  /v1/workflow/task/complete:
    post:
      consumes:
      - application/json
      description: 撤回
      parameters:
      - description: 撤回所需参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/workflow.TaskReceiver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 撤回
      tags:
      - process
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: bearer
    type: apiKey
swagger: "2.0"